refactor: subcommand admin givemoney
#145
