+++ slash_commands/subcommands/admin/index.js
@@ -1,7 +1,9 @@
 const { salon } = require("./salon");
 const { avertissement } = require("./avertissement");
+const { givemoney } = require("./givemoney");
 
 module.exports = {
     salon,
     avertissement,
+    givemoney,
 };
